requirements.txt
streamlit
openai
requests
moviepy
pydub
python-dotenv
assemblyai
elevenlabs